Geseau-master/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
blinker==1.7.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmdstanpy==1.2.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.11.0
Cython==3.0.2
Django==4.2.4
django-cors-headers==4.2.0
django-cron==0.6.0
django-filter==23.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
Flask==3.0.0
fonttools==4.42.1
frozenlist==1.4.0
holidays==0.49
idna==3.4
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.3.2
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
multidict==6.0.4
mysqlclient==2.2.0
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.1
patsy==0.5.3
Pillow==10.0.1
plotly==5.22.0
pmdarima==2.0.3
prompt-toolkit==3.0.39
prophet==1.1.5
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
requests==2.31.0
runserver==1.0.4
scikit-learn==1.3.1
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
stanio==0.5.0
statsmodels==0.14.0
tenacity==8.5.0
threadpoolctl==3.2.0
tqdm==4.66.4
twilio==8.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.7
Werkzeug==3.0.1
yarl==1.9.2
